requirements.txt
# CanonFodder Core Requirements
# This file contains the exact versions used by the core application.
# 
# NOTE: For Apache Airflow compatibility, use requirements-airflow.txt instead.
# Some packages in this file conflict with Airflow's dependencies.

acres==0.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.10
asttokens==3.0.0
attrs==24.3.0
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colormath==3.0.0
colour-science==0.4.6
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dotenv==0.9.9
Flask==3.0.3
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.1.0
feature-engine==1.8.2
featuretools==1.31.0
filelock==3.16.1
flake8==7.1.1
fonttools==4.55.3
frozenlist==1.6.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11>=0.16.0  # Security requirement - don't downgrade
hdbscan==0.8.40
holidays==0.64
httpcore>=1.0.7  # Required by httpx - don't downgrade
httplib2==0.22.0
httpx>=0.28.1  # Security requirement - don't downgrade
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipython==8.31.0
isodate==0.6.1
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2>=3.1.6  # Security requirement - don't downgrade
jiter==0.8.2
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.8
Levenshtein==0.26.1
looseversion==1.3.0
lxml==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mlxtend==0.23.3
multidict==6.4.3
musicbrainzngs==0.7.1
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
narwhals==1.41.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.1
openai>=1.60.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
patsy==1.0.1
phidata==2.7.6
pikepdf==9.5.2
pillow==11.1.0
platformdirs==4.3.6
plotly==6.1.1
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.3.1
prov==2.0.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pycodestyle==2.12.1
pycparser==2.22
# Core application uses these pydantic versions
# For Airflow compatibility, use requirements-airflow.txt instead
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==3.0.4
pyflakes==3.2.0
# Pygments 2.19.0 is used by the core application
# Airflow may require a different version (2.19.1)
Pygments==2.19.0
pylint==3.3.7
pylistenbrainz==0.5.1
PyMuPDF==1.25.2
PyMySQL==1.1.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
questionary==2.1.0
RapidFuzz==3.13.0
rdflib==6.3.2
requests~=2.32.4
rich==13.9.4
ruff==0.11.10
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
selenium==4.27.1
setuptools>=78.1.1,<81  # Security requirement - don't downgrade
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
# Core application requires SQLAlchemy 2.0 for ORM features
# For Airflow compatibility, use requirements-airflow.txt instead
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.5.0
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
windows-curses==2.4.1; platform_system == "Windows"
woodwork==0.31.0
wrapt==1.17.2
wsproto==1.2.0
xgboost==3.0.0
yarl==1.20.0

joypy~=0.2.6
folium~=0.14.0

pypopulation~=2020.3
flag~=0.1.1

branca~=0.8.1

# Showdown integration dependencies - removed due to installation issues
# These packages are optional and used only for Markdown-to-HTML conversion
# The application has fallback mechanisms and can function without them
# py-showdown>=0.1.0
# PyExecJS>=1.5.1

# WARNING: Installing apache-airflow directly from this file may cause dependency conflicts
# Use requirements-airflow.txt instead for Airflow compatibility
apache-airflow~=3.0.1
